requirements.txt
omegaconf>=2.3.0
pytorch-lightning>=1.8.0
hydra-core>=1.2.0
transformers>=4.25.0
neptune-client>=1.0.0
psutil>=5.8.0
datasets>=2.0.0
rouge-score>=0.1.2
sacrebleu>=2.0.0
wandb>=0.13.0
streamlit>=1.20.0
